+++ lunes_cms/cms/admin.py
@@ -51,30 +51,8 @@ def get_app_list(self, request):
 
 
 admin.AdminSite.get_app_list = get_app_list
-admin.site.register(
-    Discipline,
-    DisciplineAdmin,
-    list_display=(
-        "tree_actions",
-        "indented_title",
-        "released",
-        "creator_group",
-    ),
-    list_display_links=("indented_title",),
-)
-admin.site.register(
-    TrainingSet,
-    TrainingSetAdmin,
-    list_display=(
-        "tree_actions",
-        "indented_title",
-        "title",
-        "released",
-        "related_disciplines",
-        "creator_group",
-    ),
-    list_display_links=("indented_title",),
-)
+admin.site.register(Discipline, DisciplineAdmin)
+admin.site.register(TrainingSet, TrainingSetAdmin)
 admin.site.register(Document, DocumentAdmin)
 admin.site.register(GroupAPIKey, GroupAPIKeyAdmin)
 admin.site.register(Feedback, FeedbackAdmin)