+++ lunes_cms/cms/admins/discipline_admin.py
@@ -1,6 +1,10 @@
 from __future__ import absolute_import, unicode_literals
 
 from django.contrib import admin
+from django.db.models import Count, F, Q
+from django.db.models.functions import Greatest
+from django.urls import reverse
+from django.utils.safestring import mark_safe
 from django.utils.translation import ugettext_lazy as _
 
 from mptt.admin import DraggableMPTTAdmin
@@ -25,8 +29,24 @@ class DisciplineAdmin(DraggableMPTTAdmin):
     ]
     readonly_fields = ["created_by", "image_tag"]
     search_fields = ["title"]
+    list_display = [
+        "tree_actions",
+        "indented_title",
+        "released",
+        "modules_released",
+        "modules_unreleased",
+        "words_released",
+        "creator_group",
+    ]
+    list_display_links = ["indented_title"]
+    list_filter = ["released"]
     actions = ["delete_selected", "make_released", "make_unreleased"]
     list_per_page = 25
+
+    def get_list_display(self, request):
+        if request.user.is_superuser:
+            return self.list_display
+        return [display for display in self.list_display if display != "words_released"]
 
     def save_model(self, request, obj, form, change):
         """
@@ -77,10 +97,115 @@ class DisciplineAdmin(DraggableMPTTAdmin):
         :return: adjusted queryset
         :rtype: QuerySet
         """
-        qs = super(DisciplineAdmin, self).get_queryset(request)
+        qs = super().get_queryset(request)
         if request.user.is_superuser:
-            return qs.filter(creator_is_admin=True)
-        return qs.filter(created_by__in=request.user.groups.all())
+            return qs.filter(creator_is_admin=True).annotate(
+                modules_released=Count(
+                    "training_sets",
+                    filter=Q(
+                        training_sets__creator_is_admin=True,
+                        training_sets__released=True,
+                    ),
+                    distinct=True,
+                ),
+                modules_unreleased=Count(
+                    "training_sets",
+                    filter=Q(
+                        training_sets__creator_is_admin=True,
+                        training_sets__released=False,
+                    ),
+                    distinct=True,
+                ),
+                children_modules_released=Count(
+                    "children__training_sets",
+                    filter=Q(
+                        children__creator_is_admin=True,
+                        children__training_sets__creator_is_admin=True,
+                        children__training_sets__released=True,
+                    ),
+                    distinct=True,
+                ),
+                children_modules_unreleased=Count(
+                    "children__training_sets",
+                    filter=Q(
+                        children__creator_is_admin=True,
+                        children__training_sets__creator_is_admin=True,
+                        children__training_sets__released=False,
+                    ),
+                    distinct=True,
+                ),
+                words_released=Count(
+                    "training_sets__documents",
+                    filter=Q(
+                        training_sets__released=True,
+                        training_sets__documents__creator_is_admin=True,
+                        training_sets__documents__document_image__confirmed=True,
+                    ),
+                    distinct=True,
+                ),
+                children_words_released=Count(
+                    "children__training_sets__documents",
+                    filter=Q(
+                        children__creator_is_admin=True,
+                        children__training_sets__released=True,
+                        children__training_sets__documents__creator_is_admin=True,
+                        children__training_sets__documents__document_image__confirmed=True,
+                    ),
+                    distinct=True,
+                ),
+                modules_released_order=Greatest(
+                    "modules_released", "children_modules_released"
+                ),
+                modules_unreleased_order=Greatest(
+                    "modules_unreleased", "children_modules_unreleased"
+                ),
+                words_released_order=Greatest(
+                    "words_released", "children_words_released"
+                ),
+            )
+        user_groups = request.user.groups.all()
+        return qs.filter(created_by__in=user_groups).annotate(
+            modules_released=Count(
+                "training_sets",
+                filter=Q(
+                    training_sets__created_by__in=user_groups,
+                    training_sets__released=True,
+                ),
+                distinct=True,
+            ),
+            modules_unreleased=Count(
+                "training_sets",
+                filter=Q(
+                    training_sets__created_by__in=user_groups,
+                    training_sets__released=False,
+                ),
+                distinct=True,
+            ),
+            children_modules_released=Count(
+                "children__training_sets",
+                filter=Q(
+                    children__created_by__in=user_groups,
+                    children__training_sets__created_by__in=user_groups,
+                    children__training_sets__released=True,
+                ),
+                distinct=True,
+            ),
+            children_modules_unreleased=Count(
+                "children__training_sets",
+                filter=Q(
+                    children__created_by__in=user_groups,
+                    children__training_sets__created_by__in=user_groups,
+                    children__training_sets__released=False,
+                ),
+                distinct=True,
+            ),
+            modules_released_order=Greatest(
+                "modules_released", "children_modules_released"
+            ),
+            modules_unreleased_order=Greatest(
+                "modules_unreleased", "children_modules_unreleased"
+            ),
+        )
 
     def get_form(self, request, obj=None, **kwargs):
         """
@@ -145,6 +270,45 @@ class DisciplineAdmin(DraggableMPTTAdmin):
         """
         queryset.update(released=False)
 
+    @admin.display(
+        description=_("released modules"),
+        ordering="modules_released_order",
+    )
+    def modules_released(self, obj):
+        if not obj.is_leaf_node():
+            return obj.children_modules_released
+        trainingset_list = reverse("admin:cms_trainingset_changelist")
+        return mark_safe(
+            f"<a href={trainingset_list}?disciplines={obj.id}&released__exact=1>{obj.modules_released}</a>"
+        )
+
+    @admin.display(
+        description=_("unreleased modules"),
+        ordering="modules_unreleased_order",
+    )
+    def modules_unreleased(self, obj):
+        if not obj.is_leaf_node():
+            return obj.children_modules_unreleased
+        trainingset_list = reverse("admin:cms_trainingset_changelist")
+        return mark_safe(
+            f"<a href={trainingset_list}?disciplines={obj.id}&released__exact=0>{obj.modules_unreleased}</a>"
+        )
+
+    @admin.display(
+        description=_("published words in released modules"),
+        ordering="-words_released_order",
+    )
+    def words_released(self, obj):
+        if not obj.is_leaf_node():
+            return obj.children_words_released
+        document_list = reverse("admin:cms_document_changelist")
+        return mark_safe(
+            f"<a href={document_list}?disciplines={obj.id}&assigned=released&images=approved>{obj.words_released}</a>"
+        )
+
+    @admin.display(
+        description=_("creator group"),
+    )
     def creator_group(self, obj):
         """
         Include creator group of discipline in list display
@@ -161,8 +325,6 @@ class DisciplineAdmin(DraggableMPTTAdmin):
             return obj.created_by
         else:
             return None
-
-    creator_group.short_description = _("creator group")
 
     class Media:
         js = ("js/image_preview.js",)